feat: move twig renderer fully to generation of renderContent
We no longer need to care about twig renderer being used in Instruction classes or not. It will only be used where the renderContent is generated. At the moment, this means that this can only be done with the content-attribute and only if it is a string. But it should be easy to add parameter support later and increase the renderers capabilities.