feat: removes TableHeader instruction
It is going to be better to work with the correct structure (row->cell->text with explicit instructions each) for the moment removing unnecessary complexities from individual instructions and simplifying content rendering. Maybe in the future we can think about adding back in more convenience instructions, but for now let's stick to the simple base structure as much as possible.