Refactor: DataFetching
To allow for better data fetching and easier exception handling in the future, the generation of DataFetchers was changed and moved to the initialization phase of the bakery. Now they get created based on the query for the build of the RecipeProcessor. Any mistakes with placeholders or filter formats will now show up there and not halfway through the baking process.

This also adds basic support for sorting and pagination size for each query as that was still missing.

Also: both skipped tests are reactivated and running.