Doc: Adds basic documentation
Without going into technical details, this adds a basic documentation to make it easier to understand the intent of the bakery and what all the moving parts do.

It also explains some reasoning and potential problems as well as providing a list of areas to work on in the future.